% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_utils.R
\name{count_events}
\alias{count_events}
\title{Count the distinct events in a data.frame.}
\usage{
count_events(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
the number of distinct events represented in the data.frame
}
\description{
This function expects the data.frame to contain columns named 'Run',
'SubRun', and 'Event', carrying the appropriate integer values. Events with
different labels (if the data.frame is labeled) are counted as distinct.
}
