% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events}
\alias{events}
\title{art event timing data for 240 events.}
\format{
A tibble with 240 observatoins of 5 variables:
\describe{
  \item{Run}{the run number}
  \item{SubRUn}{the subrun number}
  \item{Event}{the event number}
  \item{Time}{event processing time in seconds}
  \item{sample}{sample id, an integer unique for each event}
}
}
\usage{
events
}
\description{
Several datasets containing the data recorded by the *art* framework's Timing service.
}
\details{
The event-by-event timing data written into the `TimeEvent` table, which contains one record per event.
}
\keyword{datasets}
