% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_interesting}
\alias{get_interesting}
\title{Filter a module timing data to include only those with largest median run time.}
\usage{
get_interesting(data, n)
}
\arguments{
\item{data}{a tibble, as returned by load_module_timing}

\item{n}{the number of modules labels to return. If there are ties in the
data, more will be returned.}
}
\value{
a dataframe
}
\description{
Filter a module timing data to include only those with largest median run time.
}
