% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_data.R
\name{load_module_memory_use}
\alias{load_module_memory_use}
\title{Load event-by-event module level memory use information}
\usage{
load_module_memory_use(filename)
}
\arguments{
\item{filename}{The name of the \emph{MemoryTracker} data file top open.}
}
\value{
a dataframe, with one row per module per event
}
\description{
Load event-by-event module level memory use information
}
