% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_data.R
\name{load_source_timing}
\alias{load_source_timing}
\title{Load source timing information from a \emph{TimeTracker} database}
\usage{
load_source_timing(filename, lbl = NULL)
}
\arguments{
\item{filename}{The name of the \emph{TimeTracker} database file to open.}

\item{lbl}{The label to be applied to each row in the dataframe. Default is
NULL, in which cade no label is added.}
}
\value{
a dataframe
}
\description{
Open the specified \emph{TimeTracker} database file, and read the
event-by-event timing information for the source. If \code{lbl} is
non-NULL, the returned dataframe will contain a column named \code{lbl},
containing the user-supplied label. The intent is that this label allows many
such dataframes, each with a distinct label, to be concatenated with
\code{dplyr::rbind_list} or \code{rbind}, and used for comparitive analysis.
}
\examples{
\dontrun{
d1 <- load_source_timing("woof/timing.db", "woof")
d2 <- load_source_timing("lsd6/timing.db", "lsd6")
d3 = rbind(d1, d2)
}
}
