% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_table}
\alias{load_table}
\title{Load a table from an SQLite3 database}
\usage{
load_table(filename, tablename, lbl = NULL)
}
\arguments{
\item{filename}{The name of the SQLite3 database file to open.}

\item{tablename}{The name of the table to be read.}

\item{lbl}{Optional, a label to be applied to each row in the dataframe.}
}
\value{
a dataframe
}
\description{
Open the specified SQLite database file, and read the specified table name
info a dataframe. If \code{lbl} is non-NULL, the returned dataframe will
contain a column named \code{lbl}, containing the user-supplied label. The
intent is that this label allows many such dataframes, each with a distinct
label, to be concatenated with \code{rbind}, and used for comparitive
analysis. Consult the \emph{MemoryTracker} and \emph{TimeTracker}
documentation for a list of the tables that are available.
}
\examples{
\dontrun{
d1 <- load_table("woof-memory.db", "ModuleInfo", "woof")
d2 <- load_table("cluck-memory.db", "ModuleInfo", "cluck")
mods <- rbind(d1, d2)
}
}
