% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{memory_db_tables}
\alias{memory_db_tables}
\title{memory_db_tables}
\usage{
memory_db_tables()
}
\value{
A character vector containing the names of the tables in the MemoryTracker database.
}
\description{
memory_db_tables
}
\examples{
  memory_db_tables()

}
