% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{module_bwplot}
\alias{module_bwplot}
\title{Make a box-and-whisker plot of module times from a timing dataframe.}
\usage{
module_bwplot(data, maxmodules = NA, ...)
}
\arguments{
\item{data}{The dataframe to plot, as from \code{load_module_timing}}

\item{maxmodules}{If non-NULL, plot only the top maxmodules entries from each label}

\item{...}{additional arguments passed to \code{bwplot}}
}
\value{
a lattice object, as from bwplot
}
\description{
Make a box-and-whisker plot of module times from a timing dataframe.
}
