% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_memory_use}
\alias{load_memory_use}
\title{Load memory use information from a \emph{MemoryTracker} database}
\usage{
load_memory_use(filename, tablename, lbl = NULL)
}
\arguments{
\item{filename}{The name of the \emph{MemoryTracker} data file to open.}

\item{tablename}{The name of the table to be read.}

\item{lbl}{The label to be applied to each row in the dataframe; default is
NULL.}
}
\value{
a dataframe
}
\description{
Open the specified \emph{MemoryTracker} database file, and read the specified
table name info a dataframe. If \code{lbl} is non-NULL, the returned
dataframe will contain a column named \code{lbl}, containing the
user-supplied label.  The intent is that this label allows many such
dataframes, each with a distinct label, to be concatenated with
\code{dplyr::rbind_list} or \code{rbind}, and used for comparitive analysis.
Consult the \emph{MemoryTracker} documentation for a list of the tables that
are available.
}
\examples{
\dontrun{
d1 <- load_memory_use("woof/memory.db", "ModuleInfo", "woof")
d2 <- load_memory_use("woof/memory.db", "EventInfo")
d3 <- load_memory_use("woof/memory.db", "OtherInfo")
d4 <- load_memory_use("woof/memory.db", "PeakUsage")
}
}
