% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_module_timing}
\alias{load_module_timing}
\title{Load moduline timing information from a _TimeTracker_ database}
\usage{
load_module_timing(filename, lbl)
}
\arguments{
\item{filename}{The name of the _TimeTracker_ database file to open.}

\item{lbl}{The label to be applied to each row in the dataframe.}
}
\value{
a dataframe
}
\description{
Open the specified _TimeTracker_ database file, and read the
module-by-module, event-by-event timing information. The returned dataframe
contains a column "lbl", containing a user-supplied label. The intent is
that this label allows many such dataframes, each with a distinct label,
to be concatenated with `rbind`, and used for comparitive analysis.
}
\examples{
d1 <- load_module_timing("woof/timing.db", "woof")
d2 <- load_module_timing("lsd6/timing.db", "lsd6")
d3 = rbind(d1, d2)
}

