% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_utils.R
\name{memory_db_tables}
\alias{memory_db_tables}
\title{Names of tables in a _MemoryTracker_ database.}
\usage{
memory_db_tables()
}
\value{
A character vector containing the names of the tables in the MemoryTracker database.
}
\description{
Names of tables in a _MemoryTracker_ database.
}
\examples{
  library(artsupport)
  memory_db_tables()
}
