% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{module_bwplot}
\alias{module_bwplot}
\title{Make a box-and-whisker plot of module times from a timing dataframe.}
\usage{
module_bwplot(data, maxLables = NULL, ...)
}
\arguments{
\item{data}{The dataframe to plot}
}
\value{
a lattice object, as fron bwplot
}
\description{
Make a box-and-whisker plot of module times from a timing dataframe.
}

